# Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Alfa HAK Copilot API
  description: |
    Swagger спецификация для продакшен-инстанса Alfa HAK Copilot.

    Бэкенд основан на FastAPI и обслуживает Telegram-бота, мини-веб-приложение,
    загрузку документов, калькулятор и интеграцию с Альфа-Бизнесом.
  version: "1.0.0"
  contact:
    name: Alfa HAK Copilot Team
    url: https://alpha-hak-copilot.ruka.me
servers:
  - url: https://api-alpha-hak-copilot.ruka.me/api
    description: Production (24/7)
  - url: http://localhost:8000/api
    description: Local development
...
paths:
...
  /integration/profile:
    post:
      tags: [integration]
      summary: Сохранить профиль компании из web-app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyProfile'
      responses:
        '200':
          description: Профиль сохранён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfileResponse'
  /integration/alpha-business:
    post:
      tags: [integration]
      summary: Подтвердить привязку Альфа-Бизнес
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationConfirmation'
      responses:
        '200':
          description: Интеграция активирована
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationConfirmationResponse'
  /integration/state/{user_id}:
    get:
      tags: [integration]
      summary: Получить текущий статус онбординга
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Актуальные данные по профилю и интеграции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingStateResponse'
components:
  parameters:
    UserIdPath:
      in: path
      name: user_id
      required: true
      schema:
        type: string
      description: Telegram user id
  schemas:
...
    CompanyProfile:
      type: object
      required: [user_id, company_name]
      properties:
        user_id:
          type: string
        company_name:
          type: string
        industry:
          type: string
          nullable: true
        employees:
          type: integer
          nullable: true
        annual_revenue:
          type: string
          nullable: true
        key_systems:
          type: string
          nullable: true
        goals:
          type: string
          nullable: true
        language:
          type: string
          default: ru
        created_at:
          type: string
          format: date-time
    CompanyProfileResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        profile:
          $ref: '#/components/schemas/CompanyProfile'
    IntegrationConfirmation:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
        provider:
          type: string
          default: alpha_business
        connected_at:
          type: string
          format: date-time
    IntegrationStatus:
      type: object
      properties:
        status:
          type: string
          example: connected
        provider:
          type: string
        connected_at:
          type: string
          format: date-time
          nullable: true
    IntegrationConfirmationResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        integration:
          $ref: '#/components/schemas/IntegrationStatus'
    ProfileIndexStatus:
      type: object
      properties:
        status:
          type: string
        queued_at:
          type: string
          format: date-time
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        finished_at:
          type: string
          format: date-time
          nullable: true
        reason:
          type: string
          nullable: true
    OnboardingStateResponse:
      type: object
      properties:
        user_id:
          type: string
        profile:
          $ref: '#/components/schemas/CompanyProfile'
          nullable: true
        profile_status:
          $ref: '#/components/schemas/ProfileIndexStatus'
          nullable: true
        integration:
          $ref: '#/components/schemas/IntegrationStatus'
          nullable: true